user-reg/app/requirements.txt
Django==3.2
python-decouple
djangorestframework
django-cors-headers
djangorestframework-jwt
flake8
psycopg2-binary
gunicorn
six
